backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contextlib2==21.6.0
cron_descriptor==2.0.6
cryptography==46.0.4
Django==5.2.10
django-allauth==65.14.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-ninja==1.5.3
django-ninja-extra==0.30.9
django-storages==1.14.6
django-timezone-field==7.2.1
django_celery_results==2.6.0
durationpy==0.10
fido2==2.1.1
idna==3.11
injector==0.24.0
kombu==5.6.2
kubernetes==35.0.0
oauthlib==3.3.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
websocket-client==1.9.0
croniter==2.0.1
